Added 3 validation

+++ resources/dataset.py
@@ -93,9 +93,9 @@ class MiceData(Dataset):
             img1_name = "vid{}_{}".format(vid_find,int(frame1_find))
             img2_name = "vid{}_{}".format(vid_find,int(frame2_find))
             img3_name = "vid{}_{}".format(vid_find,int(frame3_find))
-            img1 = rgb2gray(plt.imread(self.image_path+"\\"+img1_name+".jpg"))
-            img2 = rgb2gray(plt.imread(self.image_path+"\\"+img2_name+".jpg"))
-            img3 = rgb2gray(plt.imread(self.image_path+"\\"+img3_name+".jpg"))
+            img1 = rgb2gray(plt.imread(self.image_path+"\\"+img1_name+".jpg")).astype(np.float32)
+            img2 = rgb2gray(plt.imread(self.image_path+"\\"+img2_name+".jpg")).astype(np.float32)
+            img3 = rgb2gray(plt.imread(self.image_path+"\\"+img3_name+".jpg")).astype(np.float32)
             return img1,img2,img3
         
 class MiceData2(Dataset): # Pour la video 2 seulement
